prevent a public notice being created if an active one already exists

Approved-by: Chris Wood
Approved-by: Michael Heyes

+++ src/main/java/uk/co/ogauthority/pwa/service/appprocessing/publicnotice/WithdrawPublicNoticeService.java
@@ -78,8 +78,10 @@ public class WithdrawPublicNoticeService {
     var publicNotice = publicNoticeService.getLatestPublicNotice(pwaApplication);
     var publicNoticeStatusBeforeWithdrawal = publicNotice.getStatus();
 
-    var workflowTaskInstance = new WorkflowTaskInstance(publicNotice, publicNoticeStatusBeforeWithdrawal.getWorkflowTask());
-    camundaWorkflowService.deleteProcessAndTask(workflowTaskInstance);
+    if (!publicNotice.getStatus().equals(PublicNoticeStatus.PUBLISHED)) {
+      var workflowTaskInstance = new WorkflowTaskInstance(publicNotice, publicNoticeStatusBeforeWithdrawal.getWorkflowTask());
+      camundaWorkflowService.deleteProcessAndTask(workflowTaskInstance);
+    }
 
     publicNotice.setStatus(PublicNoticeStatus.WITHDRAWN);
     publicNotice.setWithdrawalReason(form.getWithdrawalReason());
